Retract MAE d_eff=63 claim; update with consistent 4-model RankMe results
Consistent RankMe (spectral entropy) on 500 EchoNet-Dynamic test videos
with scripts/rebuttal/rankme.py (HyperPod jobs 510/525):

  JEPA e95: 245.3, BYOL e100: 220.7, MAE e99: 206.4, SALT v1 e79: 202.7

All models in the 200-245 range — no 3x collapse for MAE. The prior
MAE=63 (Goodfire report) is not reproducible and should not be cited.
Effective dimensionality is demoted from primary mechanism to appendix.
The surviving mechanistic explanation is temporal structure encoding
(frame shuffling).

Also fixes rankme.py VideoMAE float32 casting and sbatch LD_LIBRARY_PATH
for H100 cuBLAS compatibility.

+++ scripts/rebuttal/rankme.py
@@ -88,7 +88,7 @@ def load_encoder(cfg, device, resolution=224, frames=16):
         state = ckpt.get("model", ckpt)
         state = _convert_pretrain_to_finetune_state_dict(state, model.state_dict())
         model.load_state_dict(state, strict=False)
-        model.eval().to(device)
+        model.float().eval().to(device)  # float32 to avoid H100 cuBLAS bf16 errors
 
         def capture_tokens(module, input, output):
             token_storage["tokens"] = output.detach()
@@ -135,11 +135,12 @@ def extract_embedding(model, clip, device, is_videomae, token_storage):
     clip = clip.to(device)
 
     if is_videomae:
+        # Force float32 to avoid H100 cuBLAS bf16 errors with VideoMAE
         token_storage.clear()
-        model.forward_features(clip)
+        model.forward_features(clip.float())
         if "tokens" not in token_storage:
             return None
-        tokens = token_storage["tokens"].squeeze(0)  # [N, D]
+        tokens = token_storage["tokens"].squeeze(0).float()  # [N, D]
     else:
         tokens = model(clip).squeeze(0)  # [N, D]
